+++ BackEnd/FJSN_ZGE_ABG_Proyecto_TFC_01/src/main/java/tiendaonline/entities/Pedido.java
@@ -33,7 +33,7 @@ public class Pedido implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_pedido")
-	private Long idPedido;
+	private Integer idPedido;
 
 	@Column(name = "fecha_pedido")
 	private LocalDate fechaPedido;
