+++ BackEnd/FJSN_ZGE_ABG_Proyecto_TFC_01/src/main/java/tiendaonline/repository/Detalle_pedidoRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tiendaonline.entities.DetallePedido;
 
-public interface Detalle_pedidoRepository extends JpaRepository<DetallePedido,Long > {
-//REVISAR TIPO
+public interface Detalle_pedidoRepository extends JpaRepository<DetallePedido,Integer > {
 }
