+++ BackEnd/FJSN_ZGE_ABG_Proyecto_TFC_01/src/main/java/tiendaonline/service/DetallePedidoService.java
@@ -12,11 +12,11 @@ public interface DetallePedidoService {
 	
 		List<DetallePedido> findAll();
 
-		DetallePedido findById(Long id);
+		DetallePedido findById(Integer id);
 
 		DetallePedido insertOne(DetallePedido detalle);
 
 		DetallePedido updateOne(DetallePedido detalle);
 
-		int deleteOne(Long id);
+		int deleteOne(Integer id);
 }
